server/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
python-whois
pydantic
requests
pillow